Fix tests for mv

+++ ffs/nix.py
@@ -309,16 +309,14 @@ def mv(resource, target):
     Return: None
     Exceptions:None
     """
-    shutil.move(resource, target)
-    # try:
-    #     shutil.move(resource, target)
-    # except shutil.Error:
-    #     if shutil._basename(resource) == target:
-    #         return # Why would you do this?
-    #     if not os.path.exists(target):
-    #         raise
-    #     rm(target, recursive=True)
-    #     shutil.move(resource, target)
+    try:
+        shutil.move(resource, target)
+    except shutil.Error:
+        if not os.path.exists(target):
+            rm(target, recursive=True)
+            shutil.move(resource, target)
+            if os.path.exists(resource):
+                rm(resource, recursive=True)
 
 
 def pwd():
